+++ packages/taskbar_integration_vincentzyu/test/taskbar_integration_test.dart
@@ -75,7 +75,6 @@ void main() {
         id: 'about',
         label: 'About',
         arguments: '--action=about',
-        separatorBefore: true,
       ),
     ];